requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
langchain-classic
langchain-huggingface
langchain-community
transformers